NewtonSolver fixes, adding tests

+++ classes/Polynomial.cpp
@@ -15,7 +15,7 @@ void Polynomial::setPolynomial(const vector<pair<int, int> > &polynomial) {
     Polynomial::polynomial = polynomial;
 }
 
-double Polynomial::value(int x) {
+double Polynomial::value(double x) {
     double value = 0;
     for (auto el : this->polynomial) {
         value += el.first * pow(x, el.second);
